# be/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.ChangePasswordRequest:
    properties:
      new_password:
        example: newpassword456
        minLength: 6
        type: string
      old_password:
        example: password123
        type: string
    required:
    - new_password
    - old_password
    type: object
...
  models.CreateUserRequest:
    properties:
      email:
        example: userbaru@example.com
        type: string
      full_name:
        example: User Baru
        type: string
      password:
        example: rahasia123
        minLength: 6
        type: string
      role_id:
        example: 10000000-0000-0000-0000-000000000004
        type: string
      username:
        example: userbaru
        maxLength: 100
        minLength: 3
        type: string
    required:
    - email
    - full_name
    - password
    - role_id
    - username
    type: object
...
  models.Permission:
    properties:
      code:
        example: user.view
        type: string
      created_at:
        type: string
      description:
        example: Can view user list and detail
        type: string
      id:
        example: a0000000-0000-0000-0000-000000000001
        type: string
      module:
        example: user
        type: string
      name:
        example: View Users
        type: string
      updated_at:
        type: string
    type: object
...
  models.Role:
    properties:
      created_at:
        type: string
      description:
        example: Administrator with management access
        type: string
      id:
        example: 10000000-0000-0000-0000-000000000002
        type: string
      is_active:
        example: true
        type: boolean
      name:
        example: admin
        type: string
      permissions:
        items:
          $ref: '#/definitions/models.Permission'
        type: array
      updated_at:
        type: string
    type: object
...
  models.UpdateUserRequest:
    properties:
      email:
        example: updated@example.com
        type: string
      full_name:
        example: Nama Baru
        type: string
      is_active:
        example: true
        type: boolean
      role_id:
        example: 10000000-0000-0000-0000-000000000004
        type: string
    type: object
  models.UserResponse:
    properties:
      email:
        example: admin@example.com
        type: string
      full_name:
        example: Administrator
        type: string
      id:
        example: 20000000-0000-0000-0000-000000000002
        type: string
      is_active:
        example: true
        type: boolean
      role:
        $ref: '#/definitions/models.Role'
      username:
        example: admin
        type: string
    type: object
  utils.APIResponse:
    properties:
      data: {}
      message:
        type: string
      meta:
        $ref: '#/definitions/utils.Meta'
      success:
        type: boolean
    type: object
  utils.Meta:
    properties:
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
host: localhost:8080
info:
  contact: {}
  description: REST API dengan autentikasi lengkap, Dynamic RBAC, dan sistem menu
    hierarki.
  title: 247 Golang API
  version: "1.0"
paths:
  /auth/change-password:
    put:
      consumes:
      - application/json
      description: Mengganti password user yang sedang login
      parameters:
      - description: Password lama dan baru
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.ChangePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Ganti password
      tags:
      - Auth
...
  /users:
    get:
      description: Mendapatkan daftar user dengan pagination. Memerlukan permission
        user.view
      parameters:
      - default: 1
        description: Nomor halaman
        in: query
        name: page
        type: integer
      - default: 10
        description: Jumlah per halaman (max 100)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.UserResponse'
                  type: array
                meta:
                  $ref: '#/definitions/utils.Meta'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Daftar semua user
      tags:
      - Users
    post:
      consumes:
      - application/json
      description: Membuat user baru dengan role tertentu. Memerlukan permission user.create
      parameters:
      - description: Data user baru
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/utils.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.UserResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Buat user baru
      tags:
      - Users
  /users/{id}:
    delete:
      description: Menghapus user berdasarkan ID. Memerlukan permission user.delete
      parameters:
      - description: User ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Hapus user
      tags:
      - Users
    get:
      description: Mendapatkan detail user berdasarkan ID. Memerlukan permission user.view
      parameters:
      - description: User ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.UserResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Detail user
      tags:
      - Users
    put:
      consumes:
      - application/json
      description: Mengupdate data user berdasarkan ID. Memerlukan permission user.update
      parameters:
      - description: User ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Data update user
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.UserResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Update user
      tags:
      - Users
securityDefinitions:
  BearerAuth:
    description: 'Masukkan token dengan format: Bearer {access_token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"